# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Ledger API
  description: Job application tracker API - Track and manage your job applications with AI-powered insights
  version: 1.0.0
  contact:
    name: API Support
    email: support@ledger.com
servers:
  - url: http://localhost:5000/api/v1
    description: Development server
  - url: https://api.ledger.com/api/v1
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /agent/invoke:
    post:
      tags:
        - Agent
      summary: Invoke AI agent for resume recommendations
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - resumeId
                - jobId
                - userGoal
              properties:
                resumeId:
                  type: string
                jobId:
                  type: string
                userGoal:
                  type: string
      responses:
        '200':
          description: Recommendations generated
...